fix: 자기이해 홈에서 Discover 결과뷰 수정 및 api 오류 수정 (#77)
